+++ hydromt/io/__init__.py
@@ -10,16 +10,9 @@ from .readers import (
     open_timeseries_from_table,
     open_vector,
     open_vector_from_table,
-    read_ini_config,
+    read_nc,
 )
-from .writers import (
-    ConfigParser,
-    configwrite,
-    netcdf_writer,
-    write_ini_config,
-    write_xy,
-    zarr_writer,
-)
+from .writers import configwrite, netcdf_writer, write_nc, write_xy, zarr_writer
 
 __all__ = [
     "open_raster",
@@ -31,11 +24,10 @@ __all__ = [
     "open_vector",
     "open_vector_from_table",
     "write_xy",
-    "write_ini_config",
     "configwrite",
     "configread",
-    "read_ini_config",
-    "ConfigParser",
     "netcdf_writer",
     "zarr_writer",
+    "read_nc",
+    "write_nc",
 ]