+++ tests/io/test_io.py
@@ -5,6 +5,7 @@ import glob
 import os
 from os.path import join
 from pathlib import Path
+from unittest.mock import MagicMock, patch
 
 import geopandas as gpd
 import numpy as np
@@ -13,6 +14,7 @@ import pytest
 import rasterio
 import xarray as xr
 from shapely.geometry import box
+from upath import UPath
 
 import hydromt
 from hydromt import _compat, raster
@@ -64,7 +66,7 @@ def test_open_vector(engine, tmpdir, df, geodf, world):
     # read shapefile
     gdf1 = hydromt.open_vector(fn_shp, bbox=list(geodf.total_bounds))
     assert np.all(gdf1 == geodf)
-    mask = gpd.GeoDataFrame({"geometry": [box(*geodf.total_bounds)]})
+    mask = gpd.GeoDataFrame({"geometry": [box(*geodf.total_bounds)]}, crs=geodf.crs)
     gdf1 = hydromt.open_vector(fn_shp, geom=mask)
     assert np.all(gdf1 == geodf)
     # read geopackage
@@ -104,13 +106,23 @@ def test_open_vector(engine, tmpdir, df, geodf, world):
         open_vector_from_table("test.fail")
 
 
+@pytest.mark.skipif(not _compat.HAS_S3FS, reason="S3FS not installed.")
+def test_open_vector_s3(geodf: gpd.GeoDataFrame):
+    m = MagicMock()
+    m.return_value = geodf
+    with patch("geopandas.io.file._read_file_fiona", m):
+        df = hydromt.open_vector(UPath("s3://fake_url/file.geojson"))
+    assert np.all(geodf == df)
+
+
 def test_open_geodataset(tmpdir, geodf):
     fn_gdf = str(tmpdir.join("points.geojson"))
     geodf.to_file(fn_gdf, driver="GeoJSON")
     # create equivalent polygon file
     fn_gdf_poly = str(tmpdir.join("polygons.geojson"))
     geodf_poly = geodf.copy()
-    geodf_poly["geometry"] = geodf_poly.buffer(0.1)
+    crs = geodf.crs
+    geodf_poly["geometry"] = geodf_poly.to_crs(3857).buffer(0.1).to_crs(crs)
     geodf_poly.to_file(fn_gdf_poly, driver="GeoJSON")
     # create zeros timeseries
     ts = pd.DataFrame(