+++ tests/metadata_resolvers/test_convention_resolver.py
@@ -1,3 +1,4 @@
+import pytest
 from polyfactory.factories.pydantic_factory import ModelFactory
 from polyfactory.pytest_plugin import register_fixture
 
@@ -13,7 +14,9 @@ class DataSourceFactory(ModelFactory[DataSource]):
     metadata_resolver = ConventionResolver()
 
 
+@pytest.mark.skip("see ticket #805 ")
 class TestConventionResolver:
+    @pytest.mark.skip(reason="test fails on windows")
     def test_resolve(self, data_source_factory: DataSourceFactory):
         mock_source = data_source_factory.build()
         resolver = ConventionResolver()